remove two lines

+++ defines.h
@@ -139,7 +139,7 @@ extern uint8_t level;
 extern uint8_t lives;
 extern uint8_t last_success_level; // to see user progress
 extern uint8_t repaint_lives;
-extern uint16_t points;
+extern uint8_t points;
 
 extern uint8_t floor_holes[][24];
